+++ project/src/pages/add-review/add-review.tsx
@@ -1,4 +1,3 @@
-import React, { useState } from 'react';
 import { Navigate, useParams } from 'react-router-dom';
 import ReviewForm from '../../components/review/review-form/review-form';
 import LogoElement from '../../components/common/logo-element/logo-element';
@@ -13,20 +12,10 @@ import { findMovieById } from '../../utils/utils';
 import useAppSelector from '../../hooks/use-app-selector/use-app-selector';
 import { getMovies } from '../../utils/selectors/selectors';
 
-type ReviewState = {
-  rating: string;
-  reviewText: string;
-}
-
 const AddReviewPage = () => {
-  const [review, setReview] = useState<ReviewState>({rating: '', reviewText: ''});
   const {id} = useParams();
   const allMovies = useAppSelector(getMovies);
-
   const currentMovie = findMovieById(allMovies, id);
-
-
-  const handleReviewChange = ({target, value}: {target: string, value: string | number}) => setReview({...review, [target]: value});
 
   if (!currentMovie) {
     return <Navigate to={AppRoute.NotFound} />;
@@ -44,7 +33,7 @@ const AddReviewPage = () => {
         </HeaderElement>
         <MoviePoster {...currentMovie} size={PosterSize.Small} />
       </div>
-      <ReviewForm handleReviewChange={handleReviewChange} />
+      <ReviewForm />
     </section>
   );
 };
